requirements.txt
# DeepFit - AI-Powered Fitness Assessment System
# Complete Requirements File

# Core Web Framework
Flask==2.3.0
Flask-CORS==4.0.0
Werkzeug==2.3.0

# Database
pymongo==4.6.0
python-dotenv==1.0.0

# Data Processing & Machine Learning
numpy==1.26.4
pandas==2.1.0
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2

# Computer Vision & AI
opencv-python==4.8.1.78
mediapipe==0.10.8
torch==2.0.1
torchvision==0.15.2

# Visualization
matplotlib==3.8.2

# Security & Authentication
bcrypt==4.0.1

# Utilities
PyYAML==6.0.1
Pillow==10.0.0
requests==2.31.0

# Optional (for enhanced features)
# ultralytics  # For YOLO models (uncomment if needed)
